# New code to perform managed <-> java lookups (typemap)