--- a/src/Xamarin.Android.Build.Tasks/Properties/Resources.resx
+++ b/src/Xamarin.Android.Build.Tasks/Properties/Resources.resx
@@ -238,6 +238,9 @@
     <value>MultiDex is enabled, but main dex list file '{0}' does not exist.</value>
     <comment>The following are literal names and should not be translated: MultiDex</comment>
   </data>
+  <data name="XA4308" xml:space="preserve">
+    <value>Failed to generate type maps</value>
+  </data>
   <data name="XA5101" xml:space="preserve">
     <value>Missing Android NDK toolchains directory '{0}'. Please install the Android NDK.</value>
     <comment>{0} - The path of the missing directory</comment>