--- a/src/Xamarin.Android.Build.Tasks/Tasks/RemoveRegisterAttribute.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/RemoveRegisterAttribute.cs
@@ -19,21 +19,17 @@ namespace Xamarin.Android.Tasks
 		[Required]
 		public ITaskItem[] ShrunkFrameworkAssemblies { get; set; }
 
-		public bool Deterministic { get; set; }
-
 		public override bool RunTask ()
 		{
 			// Find Mono.Android.dll
 			var mono_android = ShrunkFrameworkAssemblies.First (f => Path.GetFileNameWithoutExtension (f.ItemSpec) == "Mono.Android").ItemSpec;
-			var writerParameters = new WriterParameters {
-				DeterministicMvid = Deterministic,
-			};
+
 			using (var assembly = AssemblyDefinition.ReadAssembly (mono_android, new ReaderParameters { ReadWrite = true })) {
 				// Strip out [Register] attributes
 				foreach (TypeDefinition type in assembly.MainModule.Types)
 					ProcessType (type);
 
-				assembly.Write (writerParameters);
+				assembly.Write ();
 			}
 			
 			return true;