… Documentation/guides/messages/README.md
… Documentation/guides/messages/xa4308.md
… src/Mono.Android/Properties/AssemblyInfo.cs.in
… src/Mono.Android/Test/Java.Interop-Tests/Java.Interop-Tests.csproj
… src/Mono.Android/Test/Java.Interop-Tests/Properties/AssemblyInfo.cs
… src/Mono.Android/Test/Java.Interop/JnienvTest.cs
… src/Mono.Android/Test/Mono.Android-Test.Library/Mono.Android-Test.Library.csproj
… src/Mono.Android/Test/Mono.Android-Test.Library/Properties/AssemblyInfo.cs
… src/Mono.Android/Test/Mono.Android-Tests.csproj
… src/Xamarin.Android.Build.Tasks/Tasks/GenerateJavaStubs.cs
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/IncrementalBuildTest.cs
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Utilities/EnvironmentHelper.cs
… src/Xamarin.Android.Build.Tasks/Utilities/ApplicationConfigNativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblerTargetProvider.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyDataStream.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/NativeTypeMappingData.cs
… src/Xamarin.Android.Build.Tasks/Utilities/TypeMapGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/TypeMappingNativeAssemblyGenerator.cs
… src/Xamarin.Android.Build.Tasks/Utilities/X86NativeAssemblerTargetProvider.cs
… src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets
… src/monodroid/jni/android-system.cc
… src/monodroid/jni/android-system.hh
… src/monodroid/jni/application_dso_stub.cc
… src/monodroid/jni/embedded-assemblies-zip.cc
… src/monodroid/jni/embedded-assemblies.cc
… src/monodroid/jni/embedded-assemblies.hh
… src/monodroid/jni/external-api.cc
… src/monodroid/jni/monodroid-glue-internal.hh
… src/monodroid/jni/monodroid-glue.cc