@jonpryor 2020-01-15
For the eventual commit message, please provide a description of what the "file format" of the mapping code looks like, though I assume it'll resemble:

```
int JavaEntryLength;

struct ManagedToJavaEntry {
    Guid ManagedType;
    byte JavaType [JavaEntryLength];
};
struct ManagedToJavaEntries {
    int Count;
    ManagedToJavaEntry Entries [Count];
};

struct JavaToManagedEntry {
    byte JavaType [JavaEntryLength];
    Guid ManagedType;
};
struct JavaToManagedEntries {
    int Count;
    JavaToManagedEntry Entries [Count];
};
```