@grendello 2020-02-05
Please don't merge yet, I need to test fastdev mode tomorrow after the changes