@jonathanpeppers 2020-02-05
Ok, great so much better!
```
Build_CSharp_Change:
Exceeded expected time of 4100ms, actual 4160ms
Build_AndroidManifest_Change:
Exceeded expected time of 4250ms, actual 4430ms
Build_JLO_Change:
Exceeded expected time of 8150ms, actual 8630ms
```
You can bump three of these numbers:
https://github.com/xamarin/xamarin-android/blob/476921a7e526ffd522245de2bcfef93f1a91fd0a/tests/msbuild-times-reference/MSBuildDeviceIntegration.csv#L6-L10

Maybe:
```diff
-Build_CSharp_Change,4100
+Build_CSharp_Change,4200
-Build_AndroidManifest_Change,4250
+Build_AndroidManifest_Change,4500
-Build_JLO_Change,8150
+Build_JLO_Change,8700
```