@grendello 2020-02-06
OK, everything checks out locally. I applied the changes requested in https://github.com/xamarin/xamarin-android/pull/3992#issuecomment-582655425, so if this build is green-ish, I think the PR's ready for merging